Refactor datagrid row hover to hook
